Convert examples from TDB1 to TDB2
